+++ src/router/mapping.js
@@ -233,18 +233,7 @@ function getParamNames(path) {
 			if (sep.length === 2) {
 				type = sep[0];
 				if (TYPES.indexOf(type) === -1) {
-					try {
-						// parameter data type is a regex
-						// this string MUST BE a complete javascript regular expression
-						// exmaple: /^[a-zA-Z]*$/g
-						var reg = sep[0];
-						var arg = reg.substring(reg.lastIndexOf('/') + 1) || null;
-						var regStr = reg.replace(arg, '').replace(/\//g, '');
-						type = new RegExp(regStr, arg);
-					} catch (e) {
-						logger.error(e);
-						throw new Error('InvalidType: ' + type);
-					}
+					type = handleRegexDataType(type);
 				}
 				name = sep[1];
 			} else {
@@ -258,4 +247,26 @@ function getParamNames(path) {
 		}
 	}
 	return res;
+}
+
+function handleRegexDataType(type) {
+	try {
+		// parameter data type is a regex
+		// this string MUST BE a complete javascript regular expression
+		// exmaple: /^[a-zA-Z]*$/g
+		var regStr;
+		var reg = type;
+		var arg = reg.substring(reg.lastIndexOf('/') + 1) || null;
+		if (arg) {
+			regStr = reg.replace(arg, '').replace(/\//g, '');
+			type = new RegExp(regStr, arg);
+		} else {
+			regStr = reg.replace(/\//g, '');
+			type = new RegExp(regStr);
+		}
+		return type;
+	} catch (e) {
+		logger.error(e);
+		throw new Error('InvalidType: ' + type);
+	}
 }