+++ test/router/index.js
@@ -1202,4 +1202,28 @@ describe('gracenode.router', function () {
 			done();
 		});
 	});
+
+	it('can define and validate parameter type as a RegExp 2', function (done) {
+		gn.router.get('/param2/is/regex/{/(one)/:value}', function (req, res) {
+			res.json({ message: req.params.value });
+		});
+		request.GET(http + '/param/is/regex/one', {}, options, function (error, res, st) {
+			assert.equal(error, null);
+			assert.equal(st, 200);
+			assert.equal(res.message, 'one');
+			done();
+		});
+	});
+
+	it('can return 400 error for param type that is not matched by regex 2', function (done) {
+		gn.router.get('/param2/is/regex/{/(one)/:value}', function (req, res) {
+			res.json({ message: req.params.value });
+		});
+		request.GET(http + '/param/is/regex/two', {}, options, function (error, res, st) {
+			assert(error);
+			assert.equal(st, 400);
+			assert.equal(res.message, 'InvalidParameterTypeByRegExp: two');
+			done();
+		});
+	});
 });